Fixing issue with infinite caching of texture and mesh info.
Added dynamic mesh so that vertex data in a vbo can be updated rather than replaced with a new mesh.

+++ src/main/java/com/game/engine/logic/ILogic.java
@@ -5,6 +5,6 @@ public interface ILogic {
   void onStart();
   void input();
   void update();
-  void render(float fps);
+  void render(int fps);
   void onEnd();
 }
